refactor: updated routes to be isolated. index and executables file now use controller files.
