+++ .gitignore
@@ -63,3 +63,4 @@ start_container/*
 # ignore file uploads and downloads from local testing
 # public/uploads/*
 outputs/*
+src/executables/nullification/*
