+++ public/uploads/bonds/ionic_bonds.gml
@@ -1,0 +1,110 @@
+graph [
+  node [
+    id 0
+    label "27"
+    AA "LYS"
+    coord 20.959
+    coord 37.922
+    coord 50.763
+    chain "C"
+  ]
+  node [
+    id 1
+    label "39"
+    AA "ASP"
+    coord 22.973
+    coord 43.903
+    coord 49.901
+    chain "F"
+  ]
+  node [
+    id 2
+    label "59"
+    AA "ARG"
+    coord 13.451
+    coord 40.574
+    coord 57.447
+    chain "C"
+  ]
+  node [
+    id 3
+    label "35"
+    AA "ASP"
+    coord 13.536
+    coord 44.22
+    coord 52.218
+    chain "F"
+  ]
+  node [
+    id 4
+    label "76"
+    AA "GLU"
+    coord 12.07
+    coord 39.189
+    coord 61.0
+    chain "F"
+  ]
+  node [
+    id 5
+    label "83"
+    AA "ARG"
+    coord 23.398
+    coord 40.453
+    coord 47.419
+    chain "C"
+  ]
+  node [
+    id 6
+    label "87"
+    AA "ARG"
+    coord 23.833
+    coord 44.139
+    coord 46.844
+    chain "C"
+  ]
+  node [
+    id 7
+    label "102"
+    AA "HIS"
+    coord 22.798
+    coord 46.554
+    coord 49.734
+    chain "C"
+  ]
+  edge [
+    source 0
+    target 1
+    bond_type "ionic"
+    weight 4.18
+  ]
+  edge [
+    source 1
+    target 5
+    bond_type "ionic"
+    weight 4.27
+  ]
+  edge [
+    source 1
+    target 6
+    bond_type "ionic"
+    weight 3.18
+  ]
+  edge [
+    source 1
+    target 7
+    bond_type "ionic"
+    weight 2.66
+  ]
+  edge [
+    source 2
+    target 3
+    bond_type "ionic"
+    weight 4.41
+  ]
+  edge [
+    source 2
+    target 4
+    bond_type "ionic"
+    weight 4.06
+  ]
+]
