+++ src/config.js
@@ -1,0 +1,5 @@
+const task_queue_concurrency_limit = 100
+
+module.exports = {
+	task_queue_concurrency_limit
+}
