+++ src/views/layout.pug
@@ -19,7 +19,6 @@ html
       a(href='/documentation/docs/index.html') Documentation
       a(href='/upload') Upload
       a(href='/drop') Drop
-      //- a(href='/test') Test
       a(href='/executables') Executables
     
     div(data-files=data id = "fileData" )
